requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
multidict==6.0.5
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
twilio==9.1.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.1
whitenoise==6.6.0
yarl==1.9.4
